+++ pyupgw/errors.py
@@ -1,0 +1,9 @@
+"""Errors defined by the library"""
+
+
+class ClientError(Exception):
+    """Error in client operation"""
+
+
+class AuthenticationError(ClientError):
+    """Error in authentication"""